requirements.txt
asgiref==3.5.2
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
dj-database-url==1.0.0
Django==4.1
django-cors-headers==3.13.0
django-ninja==0.19.1
django-on-heroku==1.1.2
google-auth==2.10.0
google-auth-oauthlib==0.5.2
gspread==5.4.0
gspread-dataframe==3.3.0
gunicorn==20.1.0
idna==3.3
numpy==1.23.1
oauthlib==3.2.0
pandas==1.4.3
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.9.2
PyJWT==2.4.0
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.3.0
urllib3==1.26.11
whitenoise==6.2.0
